Can not show Chinese Character properly
When the Label is long Chinese String, app list only show one Chinese character and "..." (screen size: 1920x1080, dpi 480). eg. "开...".
Actually, there is no need to -3 for "...", because one Chinese Character is wide enough to show "...".
So, I think it is better to find a good enough "charToTruncate" in a loop.